ref: Split constants into new folder, create app's path constant
