+++ src/constants/path.ts
@@ -1,0 +1,7 @@
+/** Constant contains path of app */
+export const APP_PATH = {
+    DEFAULT: '/',
+    //
+    LOGIN: '/login',
+    REGISTRATION: '/registration',
+}
