+++ tsconfig.json
@@ -35,7 +35,8 @@
       "~common/*": ["src/common/*"],
       "~pages/*": ["src/pages/*"],
       "~types/*": ["src/types/*"],
-      "~img/*": ["src/img/*"]
+      "~img/*": ["src/img/*"],
+      "~constants/*": ["src/constants/*"],
     },
     "esModuleInterop": true,             /* Enables emit interoperability between CommonJS and ES Modules via creation of namespace objects for all imports. Implies 'allowSyntheticDefaultImports'. */
