+++ tslint.json
@@ -40,7 +40,7 @@
     "typedef": [true, "call-signature", "parameter"],
     "no-default-export": true,
     "match-default-export-name": false,
-    "no-implicit-dependencies": [true, ["~common", "~components", "~stores", "~pages", "~types", "~img", "~models", "~services", "~constants"]],
+    "no-implicit-dependencies": [true, ["~common", "~components", "~stores", "~pages", "~types", "~img", "~services", "~constants"]],
     "no-submodule-imports": false,
     "no-inferred-empty-object-type": false,
     "no-trailing-whitespace": [true, "ignore-jsdoc"],
